+++ src/main/java/com/mike/lunchvoter/repository/DishRepository.java
@@ -3,7 +3,7 @@ package com.mike.lunchvoter.repository;
 import com.mike.lunchvoter.entity.Dish;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DishRepository extends JpaRepository<Dish, Integer> {
+public interface DishRepository extends JpaRepository<Dish, Long> {
 
     /**
      * Checks if object with provided id exists and belongs to provided menu
@@ -12,6 +12,6 @@ public interface DishRepository extends JpaRepository<Dish, Integer> {
      * @param menuId of menu
      * @return true if such object exists in the database, false otherwise
      */
-    boolean existsByIdAndMenu_Id(Integer dishId, Integer menuId);
+    boolean existsByIdAndMenu_Id(Long dishId, Long menuId);
 
 }
