+++ src/main/java/com/mike/lunchvoter/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 
@@ -16,6 +16,6 @@ public interface UserRepository extends JpaRepository<User, Integer> {
      * @param email  of object to save
      * @return true if such object exists in the database, false otherwise
      */
-    boolean existsByIdNotAndEmailEqualsIgnoreCase(Integer userId, String email);
+    boolean existsByIdNotAndEmailEqualsIgnoreCase(Long userId, String email);
 
 }
